+++ ddsketch/store/collapsing_lowest_dense_store.go
@@ -27,7 +27,7 @@ func NewCollapsingLowestDenseStore(maxNumBins int) *CollapsingLowestDenseStore {
 }
 
 func (s *CollapsingLowestDenseStore) Add(index int) {
-	s.AddWithCount(index, float64(1))
+	s.AddWithCount(index, int32(1))
 }
 
 func (s *CollapsingLowestDenseStore) AddBin(bin Bin) {
@@ -39,7 +39,7 @@ func (s *CollapsingLowestDenseStore) AddBin(bin Bin) {
 	s.AddWithCount(index, count)
 }
 
-func (s *CollapsingLowestDenseStore) AddWithCount(index int, count float64) {
+func (s *CollapsingLowestDenseStore) AddWithCount(index int, count int32) {
 	if count == 0 {
 		return
 	}
@@ -74,7 +74,7 @@ func (s *CollapsingLowestDenseStore) extendRange(newMinIndex, newMaxIndex int) {
 	newMaxIndex = max(newMaxIndex, s.maxIndex)
 	if s.IsEmpty() {
 		initialLength := s.getNewLength(newMinIndex, newMaxIndex)
-		s.bins = make([]float64, initialLength)
+		s.bins = make([]int32, initialLength)
 		s.offset = newMinIndex
 		s.minIndex = newMinIndex
 		s.maxIndex = newMaxIndex
@@ -87,7 +87,7 @@ func (s *CollapsingLowestDenseStore) extendRange(newMinIndex, newMaxIndex int) {
 		// we may grow it before we actually reach the capacity.
 		newLength := s.getNewLength(newMinIndex, newMaxIndex)
 		if newLength > len(s.bins) {
-			tmpBins := make([]float64, newLength)
+			tmpBins := make([]int32, newLength)
 			copy(tmpBins, s.bins)
 			s.bins = tmpBins
 		}
@@ -103,7 +103,7 @@ func (s *CollapsingLowestDenseStore) adjust(newMinIndex, newMaxIndex int) {
 		newMinIndex = newMaxIndex - len(s.bins) + 1
 		if newMinIndex >= s.maxIndex {
 			// There will be only one non-empty bucket.
-			s.bins = make([]float64, len(s.bins))
+			s.bins = make([]int32, len(s.bins))
 			s.offset = newMinIndex
 			s.minIndex = newMinIndex
 			s.bins[0] = s.count
@@ -111,7 +111,7 @@ func (s *CollapsingLowestDenseStore) adjust(newMinIndex, newMaxIndex int) {
 			shift := s.offset - newMinIndex
 			if shift < 0 {
 				// Collapse the buckets.
-				n := float64(0)
+				n := int32(0)
 				for i := s.minIndex; i < newMinIndex; i++ {
 					n += s.bins[i-s.offset]
 				}
@@ -163,7 +163,7 @@ func (s *CollapsingLowestDenseStore) MergeWith(other Store) {
 }
 
 func (s *CollapsingLowestDenseStore) Copy() Store {
-	bins := make([]float64, len(s.bins))
+	bins := make([]int32, len(s.bins))
 	copy(bins, s.bins)
 	return &CollapsingLowestDenseStore{
 		DenseStore: DenseStore{
